Extract RconClient into its own module
parity.py needs to talk to Factorio over RCON without dragging in the
torch/model stack that server.py imports at top level. Move the client
verbatim to rcon.py; server.py imports it from there.
